+++ src/internal/poll/fd_poll_nacljs.go
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+// +build nacl
+
 package poll
 
 import (
